Exception Handling - delete

+++ src/main/java/com/educandoweb/course/services/UserService.java
@@ -4,17 +4,26 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
-
+import com.educandoweb.course.config.TestConfig;
 import com.educandoweb.course.entities.User;
 import com.educandoweb.course.repositories.UserRepository;
+import com.educandoweb.course.services.exceptions.DatabaseException;
 import com.educandoweb.course.services.exceptions.ResourceNotFoundException;
 
 @Service
 public class UserService {
+
+    private final TestConfig testConfig;
 	
 	@Autowired
 	private UserRepository userRepository;
+
+    UserService(TestConfig testConfig) {
+        this.testConfig = testConfig;
+    }
 	
 	public List<User> findAll(){
 		return userRepository.findAll();
@@ -30,7 +39,16 @@ public class UserService {
 	}
 	
 	public void Delete(Long id) {
-		userRepository.deleteById(id);
+		try { //Se não existir
+			if(!userRepository.existsById(id)) throw new ResourceNotFoundException(id);
+			userRepository.deleteById(id);
+		}
+		catch(ResourceNotFoundException e) {
+			throw new ResourceNotFoundException(id);
+		}
+		catch(DataIntegrityViolationException e) {
+			throw new DatabaseException(e.getMessage());
+		}
 	}
 	
 	public User update(Long id, User obj) {
